// src/app/dashboard/op-maintenance/op-maintenance-details/op-auth-plan/model/rpdma90.model.ts
export class Rpdma90 {
  m90tit1 = '';
  m90tit2 = '';
  m90lit1 = '';
  m90val1 = '';
  m90mdat = '';
  m90opos = '';
  m90pdes = '';
  m90ions = '';
  m90altc = '';
  m90li1 = '';
  m90lt1 = '';
  m90li2 = '';
  m90lt2 = '';
  m90li3 = '';
  m90lt3 = '';
  m90li4 = '';
  m90lt4 = '';
  m90li5 = '';
  m90lt5 = '';
  m90li6 = '';
  m90lt6 = '';
  m90li7 = '';
  m90lt7 = '';
  m90li8 = '';
  m90lt8 = '';
  m90li9 = '';
  m90lt9 = '';
  m90li10 = '';
  m90lt10 = '';
  m90li11 = '';
  m90lt11 = '';
  m90li12 = '';
  m90lt12 = '';
  m90li13 = '';
  m90lt13 = '';
  m90li14 = '';
  m90lt14 = '';
  m90li15 = '';
  m90lt15 = '';
  m90li16 = '';
  m90lt16 = '';
  m90li17 = '';
  m90lt17 = '';
  m90li18 = '';
  m90lt18 = '';
  m90li19 = '';
  m90lt19 = '';
  m90li20 = '';
  m90lt20 = '';
  m90li21 = '';
  m90lt21 = '';
  m90li22 = '';
  m90lt22 = '';
  m90li23 = '';
  m90lt23 = '';
  m90li24 = '';
  m90lt24 = '';
  m90li25 = '';
  m90lt25 = '';
  m90li26 = '';
  m90lt26 = '';
  m90li27 = '';
  m90lt27 = '';
  m90li28 = '';
  m90lt28 = '';
  m90li29 = '';
  m90lt29 = '';
  m90li30 = '';
  m90lt30 = '';
  m90li31 = '';
  m90lt31 = '';
  m90li32 = '';
  m90lt32 = '';
  m90li33 = '';
  m90lt33 = '';
  m90li34 = '';
  m90lt34 = '';
  m90li35 = '';
  m90lt35 = '';
  m90li36 = '';
  m90lt36 = '';
  m90li37 = '';
  m90lt37 = '';
  m90li38 = '';
  m90lt38 = '';
  m90li39 = '';
  m90lt39 = '';
  m90acpg = '';
  m90ci1 = '';
  m90cp1 = '';
  m90ci2 = '';
  m90cp2 = '';
  m90ci3 = '';
  m90cp3 = '';
  m90ci4 = '';
  m90cp4 = '';
  m90ci5 = '';
  m90cp5 = '';
  m90ci6 = '';
  m90cp6 = '';
  m90ci7 = '';
  m90cp7 = '';
  m90ci8 = '';
  m90cp8 = '';
  m90ci9 = '';
  m90cp9 = '';
  m90ci10 = '';
  m90cp10 = '';
  m90ci11 = '';
  m90cp11 = '';
  m90ci12 = '';
  m90cp12 = '';
  m90ci13 = '';
  m90cp13 = '';
  m90ci14 = '';
  m90cp14 = '';
  m90ci15 = '';
  m90cp15 = '';
  m90ci16 = '';
  m90cp16 = '';
  m90ci17 = '';
  m90cp17 = '';
  m90ci18 = '';
  m90cp18 = '';
  m90ci19 = '';
  m90cp19 = '';
  m90ci20 = '';
  m90cp20 = '';
  m90ci21 = '';
  m90cp21 = '';
  m90ci22 = '';
  m90cp22 = '';
  m90ci23 = '';
  m90cp23 = '';
  m90ci24 = '';
  m90cp24 = '';
  m90ci25 = '';
  m90cp25 = '';
  m90ci26 = '';
  m90cp26 = '';
  m90arid = '';
  m90ri1 = '';
  m90rd1 = '';
  m90ri2 = '';
  m90rd2 = '';
  m90ri3 = '';
  m90rd3 = '';
  m90ri4 = '';
  m90rd4 = '';
  m90ri5 = '';
  m90rd5 = '';
  m90ri6 = '';
  m90rd6 = '';
  m90ri7 = '';
  m90rd7 = '';
  m90ri8 = '';
  m90rd8 = '';
  m90ri9 = '';
  m90rd9 = '';
  m90ri10 = '';
  m90rd10 = '';
  m90ri11 = '';
  m90rd11 = '';
  m90ri12 = '';
  m90rd12 = '';
  m90ri13 = '';
  m90rd13 = '';
  m90ri14 = '';
  m90rd14 = '';
  m90ri15 = '';
  m90rd15 = '';
  m90ri16 = '';
  m90rd16 = '';
  m90ri17 = '';
  m90rd17 = '';
  m90ri18 = '';
  m90rd18 = '';
  m90ri19 = '';
  m90rd19 = '';
  m90ri20 = '';
  m90rd20 = '';
  m90ri21 = '';
  m90rd21 = '';
  m90ri22 = '';
  m90rd22 = '';
  m90ri23 = '';
  m90rd23 = '';
  m90ri24 = '';
  m90rd24 = '';
  m90ri25 = '';
  m90rd25 = '';
  m90ri26 = '';
  m90rd26 = '';
  m90ri27 = '';
  m90rd27 = '';
  m90ri28 = '';
  m90rd28 = '';
  m90ri29 = '';
  m90rd29 = '';
  m90ri30 = '';
  m90rd30 = '';
  m90ri31 = '';
  m90rd31 = '';
  m90ri32 = '';
  m90rd32 = '';
  m90ri33 = '';
  m90rd33 = '';
  m90ri34 = '';
  m90rd34 = '';
  m90ri35 = '';
  m90rd35 = '';
  m90ri36 = '';
  m90rd36 = '';
  m90ri37 = '';
  m90rd37 = '';
  m90ri38 = '';
  m90rd38 = '';
  m90ri39 = '';
  m90rd39 = '';
  m90err = '';
  m90more = '';
  m90pf5 = '';
  map90Title1 = '';
  map90Title2 = '';
  map90Literal = '';
  map90Value = '';
  map90MaintDate = '';
  map90Position = '';
  map90PosDesc = '';
  map90Ions = '';
  map90LtcAll = '';
  map90LtcInfos: Map90LtcInfo[] = [];
  map90CpgAll = '';
  map90CpgInfos: Map90CpgInfo[] = [];
  map90RiderAll = '';
  map90RiderInfos: Map90RiderInfo[] = [];
  map90Err1 = '';
  map90MoreMsg = '';
  map90Pf5Lit = '';
  currentPage = '1';
  ltcList: [] = [];
  cpgList: [] = [];
  riderList: [] = [];
}